no_std support in the Rust guest bindings.
This adds no_std support in the Rust guest bindings. Mostly this
involves using `core` and `alloc` instead of `std`, but it also
involves adding `extern crate alloc;` in a few places, and also
adding a `"std"` cargo feature to gen-guest-rust so that the
`impl std::error::Error` can be made conditional.

This will eventually be useful for generating bindings from the
WASI wit files for std itself to use. And, it's useful for
experimenting with generating minimal bindings.