Add a `raw_strings` option.
Add a `raw_strings` option for the Rust guest bindings, to use `&[u8]`
and `Vec<u8>` instead of `&str` and `String` for representing strings
in bindings. This will allow polyfill modules that don't already have
a `&str` to avoid doing a `str::from_utf8` to obtain one, which is
redundant since the adapter will already be doing UTF-8 validation.