implement support for resources for C and Rust generators
This also provides stubs for the Java, Go, and Markdown generators.

In addition to resource support, I've changed how the Rust guest generator
handles exports per a conversation with Alex.  Instead of generating polymorphic
functions plus an `export_{name}!` to monomorphize them, we now require trait
implementations to be specified as parameters to the `wit_bindgen::generate!`
macro.  This allows us to avoid complicated use of generics when exporting
resources.

For example:

```
// in foo.wit:
world foo {
    export a: func()
    export b: interface {
        c: func()
        resource d {
            e: func()
        }
    }
}
```

```
wit_bindgen::generate!({
    path: "foo.wit",
    world_exports: MyFoo,
    interface_exports: {
         "b": MyB
    }
    resource_exports: {
        "b::d": MyD
    }
});

struct MyFoo;

impl Foo for MyFoo {
    fn a() { ... }
}

struct MyB;

impl exports::b::B for MyB {
    fn c() { ... }
}

struct MyD;

impl exports::b::D for MyD {
    fn e(&self) { ... }
}
```

One side-effect of the above change is that the generator needs to create stub
implementations to make the codegen tests work, plus we have to be careful to
avoid name conflicts among top-level exports (hence renaming the top level
exports of lift-lower-foreign.wit to avoid clashes with issue573.wit).

Note that there are no runtime tests yet since the Wasmtime implementation of
resources is not complete as of this writing.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>