Optimize away the `cabi_realloc` call in stream read functions.
The wasi-io input-stream function `read()` returns a `list<u8>` by
value. In practice, that means that it has to call `cabi_realloc` on
every call to allocate memory for the list, and then copy the data
from that memory into the user's buffer.

This patch adds an option to create specialized Rust and C bindings
for `read` and similar functions, where the user passes in a buffer,
and it gets used to hold the list, so that no `cabi_realloc` call is
needed.